Move NASA7 helper functions to base testing class
So we can reuse them in the vector version.
